将 WEB 接口服务器中冒险者中介的处理代码重新归整
