+++ src/web/web-server.vcxproj
@@ -193,7 +193,6 @@
     <ClCompile Include="charconfig_controller.cpp" />
     <ClCompile Include="emblem_controller.cpp" />
     <ClCompile Include="merchantstore_controller.cpp" />
-    <ClCompile Include="recruitment_controller.cpp" />
     <ClCompile Include="partybooking_controller.cpp" />
     <ClCompile Include="sqllock.cpp" />
     <ClCompile Include="userconfig_controller.cpp" />
@@ -206,7 +205,6 @@
     <ClInclude Include="emblem_controller.hpp" />
     <ClInclude Include="http.hpp" />
     <ClInclude Include="merchantstore_controller.hpp" />
-    <ClInclude Include="recruitment_controller.hpp" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="partybooking_controller.hpp" />
     <ClInclude Include="sqllock.hpp" />
