+++ src/web/web-server.vcxproj.filters
@@ -42,9 +42,6 @@
     <ClCompile Include="webutils.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="recruitment_controller.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="auth.hpp">
@@ -78,9 +75,6 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="resource.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="recruitment_controller.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="webutils.hpp">
