+++ src/web/web.hpp
@@ -90,9 +90,6 @@ extern char login_table[32];
 extern char guild_emblems_table[32];
 extern char user_configs_table[32];
 extern char char_configs_table[32];
-#ifdef Pandas_WebServer_Implement_PartyRecruitment
-extern char recruitment_table[32];
-#endif // Pandas_WebServer_Implement_PartyRecruitment
 extern char guild_db_table[32];
 extern char char_db_table[32];
 extern char merchant_configs_table[32];
